Support INT8 VIT from ONNX
    Assign HiFi4 for INT output dtype when creating fork nop

    Support commute transpose through squeeze

    Verify VIT end to end

    (cherry picked from commit 7244b5c464d93f6e7c5ab28cddabccc79c33b287)

+++ pybuda/pybuda/op/eval/buda/quantize.py
@@ -34,8 +34,8 @@ def eval(type, attr, ops):
         
     elif type == "dequantization":
         zero_point, axis = attr
-        input_float = ops[0]
-        scale = ops[1]
+        input_float = ops[0].float()
+        scale = ops[1].float()
         output_float = (input_float - zero_point) * scale
         return output_float
